feat: esm & cjs compatibility
