+++ package.json
@@ -1,35 +1,37 @@
 {
   "name": "uicons.js",
-  "version": "1.1.4",
+  "version": "1.2.0",
   "description": "UICONS JavaScript Library",
   "repository": "https://github.com/TurtIeSocks/uicons.js",
   "author": "TurtIeSocks <58572875+TurtIeSocks@users.noreply.github.com>",
   "homepage": "https://TurtIeSocks.github.io/uicons.js/",
   "license": "MIT",
   "private": false,
+  "type": "module",
   "files": [
     "dist"
   ],
-  "main": "dist/index.js",
+  "main": "dist/index.cjs",
   "types": "dist/index.d.ts",
+  "module": "dist/index.mjs",
   "scripts": {
     "start": "vite",
     "build": "vite build -- -r",
-    "build:ts": "tsc --project tsconfig.build.json"
+    "build:ts": "vite build"
   },
   "devDependencies": {
-    "@na-ji/pogo-protos": "latest",
-    "@rollup/plugin-typescript": "^11.1.6",
-    "@types/node": "^20.11.26",
-    "@types/react": "^18.2.65",
-    "@types/react-dom": "^18.2.22",
-    "@vitejs/plugin-react": "^4.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "@na-ji/pogo-protos": "^2",
+    "@rollup/plugin-typescript": "11.1.6",
+    "@types/node": "22.4.1",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
+    "@vitejs/plugin-react-swc": "3.7.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "react-virtuoso": "^4.7.2",
     "tslib": "^2.6.2",
-    "typescript": "^5.4.2",
-    "vite": "^5.1.6",
-    "vite-plugin-checker": "^0.6.4"
+    "typescript": "5.5.4",
+    "vite": "5.4.1",
+    "vite-plugin-checker": "0.7.2"
   }
 }
